+++ tinymce/tests/test_widgets.py
@@ -1,7 +1,9 @@
 # coding: utf-8
 
+from django import forms
 from django.test import TestCase
 from django.test.utils import override_settings
+from django.utils.translation import override
 
 from tinymce.widgets import get_language_config, TinyMCE
 
@@ -18,6 +20,10 @@ class TestWidgets(TestCase):
             'spellchecker_rpc_url': '/tinymce/spellchecker/'
         }
         self.assertEqual(config, config_ok)
+        with override(None):
+            # Even when no language is activated
+            config = get_language_config()
+            self.assertEqual(config, config_ok)
 
     @override_settings(LANGUAGES_BIDI=['en'])
     def test_default_config_rtl(self):
@@ -49,3 +55,19 @@ class TestWidgets(TestCase):
         self.assertIn('name="foobar"', html)
         self.assertIn('lorem ipsum', html)
         self.assertIn('class="tinymce"', html)
+        html = widget.render(
+            'foobar', 'lorem ipsum', attrs={'id': 'id_foobar', 'class': 'foo'}
+        )
+        self.assertIn('class="foo tinymce"', html)
+
+    def test_tinymce_widget_required(self):
+        """
+        The TinyMCE widget should never output the required HTML attribute, even
+        if the form field is required, as the client-side browser validation
+        might prevent form submission.
+        """
+        class TinyForm(forms.Form):
+            field = forms.CharField(required=True, widget=TinyMCE())
+
+        rendered = str(TinyForm())
+        self.assertNotIn('required', rendered)